Prevent Slack spamming
Re-wrote script to stop spamming Slack with notifications. Have used the internal pull-up resistor GPIO setting to trigger the command, as we only need to know if the soil is dry. Have also changed the sleep times to check less frequently (every 15 minutes if the soil is moist and every 45 minutes if the soil is dry).

Also moved the GPIO setup commands to the beginning of the script for convenience.